Refactor application settings module, add splash screen.
Add simple splash screen (with placeholder image) which displays on launch.
Change application window display ordering, the old 'splash' dialog (now RecentProjectsDialog) is not displayed unless there is no saved project for the user in their local settings, and it is spawned from the MainWindow (just before the MainWindow is set to visible)
Application settings (QSettings) methods and classes moved into their own module (gui/settings.py), while some of the objects are re-exported via gui/__init__.
